backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
azure-cosmos==4.5.1
azure-ai-textanalytics==5.3.0
openai==1.3.7
httpx==0.25.2
redis==5.0.1
opencensus-ext-azure==1.1.11
opencensus-ext-logging==0.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6 